Use game-specific generation tab
