+++ randovania/games/dread/gui/preset_settings/__init__.py
@@ -7,8 +7,8 @@ from randovania.interface_common.preset_editor import PresetEditor
 def dread_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
     from randovania.gui.preset_settings.trick_level_tab import PresetTrickLevel
     from randovania.gui.preset_settings.starting_area_tab import PresetMetroidStartingArea
-    from randovania.gui.preset_settings.generation_tab import PresetGeneration
     from randovania.gui.preset_settings.location_pool_tab import PresetLocationPool
+    from randovania.games.dread.gui.preset_settings.dread_generation_tab import PresetDreadGeneration
     from randovania.games.dread.gui.preset_settings.dread_patches_tab import PresetDreadPatches
     from randovania.games.dread.gui.preset_settings.dread_item_pool_tab import DreadPresetItemPool
     from randovania.games.dread.gui.preset_settings.dread_energy_tab import PresetDreadEnergy
@@ -19,7 +19,7 @@ def dread_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
         *([
               PresetMetroidStartingArea,
           ] if window_manager.is_preview_mode else []),
-        PresetGeneration,
+        PresetDreadGeneration,
         PresetLocationPool,
         PresetDreadGoal,
         DreadPresetItemPool,
