+++ randovania/games/dread/json_data/header.json
@@ -1199,7 +1199,7 @@
                 "extra": {}
             },
             "HighDanger": {
-                "long_name": "Allow Highly Dangerous Logic",
+                "long_name": "Highly Dangerous Logic",
                 "extra": {}
             }
         },
